Add all file manipulation related methods to file-handler.js
Add new openFile and showOpenModal methods and move saveFile and showSaveModal from save-file.js to file-handler.js. Replace calls in action files.
